backend/requirements.txt
aioredis==2.0.1
alembic==1.13.1
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
ariadne==0.22
async-timeout==4.0.3
asyncpg==0.29.0
billiard==4.2.1
celery==5.4.0
certifi==2024.2.2
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
dnspython==2.6.1
email-validator==2.1.0.post1
environs==11.0.0
fastapi==0.109.2
fastapi-cache2==0.2.2
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.4.2
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.23.0
nodeenv==1.9.1
orjson==3.9.14
packaging==23.2
passlib==1.7.4
pendulum==3.0.0
pluggy==1.4.0
prompt_toolkit==3.0.48
pycparser==2.22
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.16.2
PyJWT==2.9.0
pyright==1.1.385
pytest==8.0.1
pytest-asyncio==0.23.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.1.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
time-machine==2.16.0
typing_extensions==4.9.0
tzdata==2024.2
ujson==5.9.0
uvicorn==0.27.1
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0

